Kill idle interactive sessions to allow worker nodes to be released.
Add xidlekill bash script, which watches the session idle time. If a
preconfigured limit is exceeded, a message is displayed on the desktop;
if the user does not react within a preconfigured interval, a kill
request (with -TERM) is sent to the pool user's processes.

The puppet configuration is modified to include the new script, and to
add
a package requirement for xprintidle (which is used in the script).

The prepareaccount script is modified to add xidlekill to the autostart
script.
MachineEJB is modified to read the two preconfigured timeout values from
the properties file and pass them to the prepareaccount script.  The
example
properties script is modified to add the new properties (though default
values
are used if the properties are not defined).