Comment on puppet modules location
For possible future reference. When attempting an installation under
RHEL and a more recent release of Puppet, it transpired that the
location of the puppet modules had changed, so that the hard-wired value
of USERGEN was no longer correct. Changing it solved the immediate
problem; but other problems have so far discouraged further attempts to
get a RHEL installation to work.